Moved all table view cells to one class
